refactor the test database client to be a simple mock and then add to it in a specific test, in this case the get_all_sources test
